Almost Final Polish
Had to revert back to an old version, The items now show up in the inventory GUI and can be used when they need to be used. Flashlight works how it should, all the assets have been re-added for the most part. It does not look like we will get the book to work though and that was a big mechanic.